# myhomebp-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyHomeBP API",
        "description": "Blood Pressure Management API for NHS and Private GP Surgery Use in the UK",
        "contact": {
            "name": "247 Meditech Support",
            "email": "support@247meditech.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.myhomebp.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/blood-pressure/record": {
            "post": {
                "tags": [
                    "Blood Pressure"
                ],
                "summary": "Record blood pressure reading",
                "operationId": "6eb0fca4b8482d36fd3be1148759b816",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reading_date",
                                    "session_type",
                                    "reading_1_systolic",
                                    "reading_1_diastolic",
                                    "reading_1_pulse",
                                    "reading_2_systolic",
                                    "reading_2_diastolic",
                                    "reading_2_pulse"
                                ],
                                "properties": {
                                    "reading_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-01-15T09:30:00Z"
                                    },
                                    "session_type": {
                                        "type": "string",
                                        "enum": [
                                            "am",
                                            "pm"
                                        ],
                                        "example": "am"
                                    },
                                    "reading_1_systolic": {
                                        "type": "integer",
                                        "example": 120
                                    },
                                    "reading_1_diastolic": {
                                        "type": "integer",
                                        "example": 80
                                    },
                                    "reading_1_pulse": {
                                        "type": "integer",
                                        "example": 72
                                    },
                                    "reading_2_systolic": {
                                        "type": "integer",
                                        "example": 118
                                    },
                                    "reading_2_diastolic": {
                                        "type": "integer",
                                        "example": 78
                                    },
                                    "reading_2_pulse": {
                                        "type": "integer",
                                        "example": 70
                                    },
                                    "reading_3_systolic": {
                                        "type": "integer",
                                        "example": 115
                                    },
                                    "reading_3_diastolic": {
                                        "type": "integer",
                                        "example": 75
                                    },
                                    "reading_3_pulse": {
                                        "type": "integer",
                                        "example": 68
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Blood pressure reading recorded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Blood pressure reading recorded successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "reading": {
                                                    "type": "object"
                                                },
                                                "system_response": {
                                                    "type": "string",
                                                    "example": "Thank you for submitting today's reading."
                                                },
                                                "requires_third_reading": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/blood-pressure/readings": {
            "get": {
                "tags": [
                    "Blood Pressure"
                ],
                "summary": "Get blood pressure readings",
                "operationId": "f54671d9a14c2476e529a5600b64ba33",
                "parameters": [
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Number of days to retrieve readings for",
                        "schema": {
                            "type": "integer",
                            "example": 7
                        }
                    },
                    {
                        "name": "session_type",
                        "in": "query",
                        "description": "Filter by session type",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "am",
                                "pm"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blood pressure readings retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "readings": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "seven_day_average": {
                                                    "type": "object"
                                                },
                                                "am_average": {
                                                    "type": "object"
                                                },
                                                "pm_average": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/blood-pressure/averages": {
            "get": {
                "tags": [
                    "Blood Pressure"
                ],
                "summary": "Get blood pressure averages and trends",
                "operationId": "43a174b8eeb02a831538c8dc6f4fc69c",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Time period for averages",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "7",
                                "30",
                                "90"
                            ],
                            "example": "7"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blood pressure averages retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "overall_average": {
                                                    "type": "object"
                                                },
                                                "am_average": {
                                                    "type": "object"
                                                },
                                                "pm_average": {
                                                    "type": "object"
                                                },
                                                "trend_data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the token with the `Bearer: ` prefix, e.g. `Bearer abc123`",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}